Central housing bottom plate.
In this simple implementation, the bottom plate is automatically shared
between the central housing, its adapter and the rest of the case on
one side, but cuts off at the origin. Also, the preview of the bottom
plate displayed with the central housing does not take wrist-rest
combination settings into account.