Separate tweak parameters.
Added a separate parameter for tweaking the shape of the central
housing. Like the last commit, this is not a pretty solution. For
example, it will not affect the shape of the bottom plate, and the
code should probably be organized into a new module and use filtering
as for bottom-plate fasteners, rather than touching the access
module.