MCU and port types. Shelf. Sides.
Added new MCU types.

Added ports, replacing a less flexible connection configuration
and broadening support for common connectors. In this commit,
negative tweaks are not yet supported, so holders have no internal
aperture for wiring, and holder aliases are not used.

Added an MCU shelf as a type of support, implementing this for the
Dactyl-ManuForm in a manner mimicking the original `teensy-holder`.

In anchoring to features, transitioned from corners to a broader
concept of sides. Added more segments for the rear housing and
removed some conveniences specific to the rear housing.

Fixed a bug that made shallow LED channels, and did a bit of
linting and other chores.