Richer placement.
Reworked placement in relation to named features, with the aim of a more
general fix for the problem of offsets being applied twice and/or in
type-specific contexts. As part of this work, mcu-pcba is now a built-in
anchor, anchoring “offset” has been split into intrinsic and extrinsic
forms, rotation has been moved into anchoring, and a
“preserve-orientation” parameter has been added to allow more control.

2D-specific forms of anchoring have been removed from the configuration
layer and are now implemented in the application layer.

Two new parameter schema modules were added because anchoring needed to
be defined more centrally for use in multiple parsers, including
arbitrary shapes.

Added a new document on anchoring, splitting this off from the general
configuration guide to live alongside the document on arbitrary shapes.

Removed foot plates as a feature. These are now obviated by richer
tweaking.

Also fixed an unrelated problem drawing lines between above-ground
points of the central housing interface, in cases where an intermediate
point is below ground.

Committed with known error rendering Concertina (central housing adapter
in main body).