Improved previews.
Made it possible to include (that is, pre-empt the exclusion of) custom
bodies for visualization. config/visualization.yaml now names all custom
bodies in the Concertina for this purpose, but a “make vis
concertina_64key” job won’t include those bodies in previews, because
Make does not put the modified file last. It’s necessary to run the
program directly.

There is still no automated preview of a bilaterally complete
Concertina.