Bottom flanges, start.
Major rebuild of bottom plate configuration with breaking changes.
This replaces the rebuild of bottom plates already in progress; it is no
longer true that each position in a plate gets a unique name.

This commit removes and does not fully replace projections for
central-housing bottom-plate bosses. The Concertina has temporarily lost
some of its bosses completely, and some others are broken. DFM is also
not fully maintained in this commit. All will be fixed later.

The verbosity of changes in e.g. config/base.yaml needs a solution but
that may come later still.

Upgraded to scad-app v1.0.0, which was needed to fix problems along
the way. A few unrelated passages of files were rearranged for
alphabetical or pedagogical order.